[GH_PR] support r/w floating(single) value type (CON-1077)

See merge request app-frameworks/esp-matter!674

+++ components/esp_matter/esp_matter_attribute.h
@@ -479,14 +479,14 @@ attribute_t *create_replacement_product_list(cluster_t *cluster, uint8_t *value,
 
 namespace carbon_monoxide_concentration_measurement {
 namespace attribute {
-attribute_t *create_measured_value(cluster_t *cluster, nullable<uint16_t> value);
-attribute_t *create_min_measured_value(cluster_t *cluster, nullable<uint16_t> value);
-attribute_t *create_max_measured_value(cluster_t *cluster, nullable<uint16_t> value);
-attribute_t *create_peak_measured_value(cluster_t *cluster, nullable<uint16_t> value);
+attribute_t *create_measured_value(cluster_t *cluster, nullable<float> value);
+attribute_t *create_min_measured_value(cluster_t *cluster, nullable<float> value);
+attribute_t *create_max_measured_value(cluster_t *cluster, nullable<float> value);
+attribute_t *create_peak_measured_value(cluster_t *cluster, nullable<float> value);
 attribute_t *create_peak_measured_value_window(cluster_t *cluster, uint32_t value);
-attribute_t *create_average_measured_value(cluster_t *cluster, nullable<uint16_t> value);
+attribute_t *create_average_measured_value(cluster_t *cluster, nullable<float> value);
 attribute_t *create_average_measured_value_window(cluster_t *cluster, uint32_t value);
-attribute_t *create_uncertainty(cluster_t *cluster, uint16_t value);
+attribute_t *create_uncertainty(cluster_t *cluster, float value);
 attribute_t *create_measurement_unit(cluster_t *cluster, uint8_t value);
 attribute_t *create_measurement_medium(cluster_t *cluster, uint8_t value);
 attribute_t *create_level_value(cluster_t *cluster, uint8_t value);
@@ -495,14 +495,14 @@ attribute_t *create_level_value(cluster_t *cluster, uint8_t value);
 
 namespace carbon_dioxide_concentration_measurement {
 namespace attribute {
-attribute_t *create_measured_value(cluster_t *cluster, nullable<uint16_t> value);
-attribute_t *create_min_measured_value(cluster_t *cluster, nullable<uint16_t> value);
-attribute_t *create_max_measured_value(cluster_t *cluster, nullable<uint16_t> value);
-attribute_t *create_peak_measured_value(cluster_t *cluster, nullable<uint16_t> value);
+attribute_t *create_measured_value(cluster_t *cluster, nullable<float> value);
+attribute_t *create_min_measured_value(cluster_t *cluster, nullable<float> value);
+attribute_t *create_max_measured_value(cluster_t *cluster, nullable<float> value);
+attribute_t *create_peak_measured_value(cluster_t *cluster, nullable<float> value);
 attribute_t *create_peak_measured_value_window(cluster_t *cluster, uint32_t value);
-attribute_t *create_average_measured_value(cluster_t *cluster, nullable<uint16_t> value);
+attribute_t *create_average_measured_value(cluster_t *cluster, nullable<float> value);
 attribute_t *create_average_measured_value_window(cluster_t *cluster, uint32_t value);
-attribute_t *create_uncertainty(cluster_t *cluster, uint16_t value);
+attribute_t *create_uncertainty(cluster_t *cluster, float value);
 attribute_t *create_measurement_unit(cluster_t *cluster, uint8_t value);
 attribute_t *create_measurement_medium(cluster_t *cluster, uint8_t value);
 attribute_t *create_level_value(cluster_t *cluster, uint8_t value);
@@ -511,14 +511,14 @@ attribute_t *create_level_value(cluster_t *cluster, uint8_t value);
 
 namespace nitrogen_dioxide_concentration_measurement {
 namespace attribute {
-attribute_t *create_measured_value(cluster_t *cluster, nullable<uint16_t> value);
-attribute_t *create_min_measured_value(cluster_t *cluster, nullable<uint16_t> value);
-attribute_t *create_max_measured_value(cluster_t *cluster, nullable<uint16_t> value);
-attribute_t *create_peak_measured_value(cluster_t *cluster, nullable<uint16_t> value);
+attribute_t *create_measured_value(cluster_t *cluster, nullable<float> value);
+attribute_t *create_min_measured_value(cluster_t *cluster, nullable<float> value);
+attribute_t *create_max_measured_value(cluster_t *cluster, nullable<float> value);
+attribute_t *create_peak_measured_value(cluster_t *cluster, nullable<float> value);
 attribute_t *create_peak_measured_value_window(cluster_t *cluster, uint32_t value);
-attribute_t *create_average_measured_value(cluster_t *cluster, nullable<uint16_t> value);
+attribute_t *create_average_measured_value(cluster_t *cluster, nullable<float> value);
 attribute_t *create_average_measured_value_window(cluster_t *cluster, uint32_t value);
-attribute_t *create_uncertainty(cluster_t *cluster, uint16_t value);
+attribute_t *create_uncertainty(cluster_t *cluster, float value);
 attribute_t *create_measurement_unit(cluster_t *cluster, uint8_t value);
 attribute_t *create_measurement_medium(cluster_t *cluster, uint8_t value);
 attribute_t *create_level_value(cluster_t *cluster, uint8_t value);
@@ -527,14 +527,14 @@ attribute_t *create_level_value(cluster_t *cluster, uint8_t value);
 
 namespace ozone_concentration_measurement {
 namespace attribute {
-attribute_t *create_measured_value(cluster_t *cluster, nullable<uint16_t> value);
-attribute_t *create_min_measured_value(cluster_t *cluster, nullable<uint16_t> value);
-attribute_t *create_max_measured_value(cluster_t *cluster, nullable<uint16_t> value);
-attribute_t *create_peak_measured_value(cluster_t *cluster, nullable<uint16_t> value);
+attribute_t *create_measured_value(cluster_t *cluster, nullable<float> value);
+attribute_t *create_min_measured_value(cluster_t *cluster, nullable<float> value);
+attribute_t *create_max_measured_value(cluster_t *cluster, nullable<float> value);
+attribute_t *create_peak_measured_value(cluster_t *cluster, nullable<float> value);
 attribute_t *create_peak_measured_value_window(cluster_t *cluster, uint32_t value);
-attribute_t *create_average_measured_value(cluster_t *cluster, nullable<uint16_t> value);
+attribute_t *create_average_measured_value(cluster_t *cluster, nullable<float> value);
 attribute_t *create_average_measured_value_window(cluster_t *cluster, uint32_t value);
-attribute_t *create_uncertainty(cluster_t *cluster, uint16_t value);
+attribute_t *create_uncertainty(cluster_t *cluster, float value);
 attribute_t *create_measurement_unit(cluster_t *cluster, uint8_t value);
 attribute_t *create_measurement_medium(cluster_t *cluster, uint8_t value);
 attribute_t *create_level_value(cluster_t *cluster, uint8_t value);
@@ -543,14 +543,14 @@ attribute_t *create_level_value(cluster_t *cluster, uint8_t value);
 
 namespace formaldehyde_concentration_measurement {
 namespace attribute {
-attribute_t *create_measured_value(cluster_t *cluster, nullable<uint16_t> value);
-attribute_t *create_min_measured_value(cluster_t *cluster, nullable<uint16_t> value);
-attribute_t *create_max_measured_value(cluster_t *cluster, nullable<uint16_t> value);
-attribute_t *create_peak_measured_value(cluster_t *cluster, nullable<uint16_t> value);
+attribute_t *create_measured_value(cluster_t *cluster, nullable<float> value);
+attribute_t *create_min_measured_value(cluster_t *cluster, nullable<float> value);
+attribute_t *create_max_measured_value(cluster_t *cluster, nullable<float> value);
+attribute_t *create_peak_measured_value(cluster_t *cluster, nullable<float> value);
 attribute_t *create_peak_measured_value_window(cluster_t *cluster, uint32_t value);
-attribute_t *create_average_measured_value(cluster_t *cluster, nullable<uint16_t> value);
+attribute_t *create_average_measured_value(cluster_t *cluster, nullable<float> value);
 attribute_t *create_average_measured_value_window(cluster_t *cluster, uint32_t value);
-attribute_t *create_uncertainty(cluster_t *cluster, uint16_t value);
+attribute_t *create_uncertainty(cluster_t *cluster, float value);
 attribute_t *create_measurement_unit(cluster_t *cluster, uint8_t value);
 attribute_t *create_measurement_medium(cluster_t *cluster, uint8_t value);
 attribute_t *create_level_value(cluster_t *cluster, uint8_t value);
@@ -559,14 +559,14 @@ attribute_t *create_level_value(cluster_t *cluster, uint8_t value);
 
 namespace pm1_concentration_measurement {
 namespace attribute {
-attribute_t *create_measured_value(cluster_t *cluster, nullable<uint16_t> value);
-attribute_t *create_min_measured_value(cluster_t *cluster, nullable<uint16_t> value);
-attribute_t *create_max_measured_value(cluster_t *cluster, nullable<uint16_t> value);
-attribute_t *create_peak_measured_value(cluster_t *cluster, nullable<uint16_t> value);
+attribute_t *create_measured_value(cluster_t *cluster, nullable<float> value);
+attribute_t *create_min_measured_value(cluster_t *cluster, nullable<float> value);
+attribute_t *create_max_measured_value(cluster_t *cluster, nullable<float> value);
+attribute_t *create_peak_measured_value(cluster_t *cluster, nullable<float> value);
 attribute_t *create_peak_measured_value_window(cluster_t *cluster, uint32_t value);
-attribute_t *create_average_measured_value(cluster_t *cluster, nullable<uint16_t> value);
+attribute_t *create_average_measured_value(cluster_t *cluster, nullable<float> value);
 attribute_t *create_average_measured_value_window(cluster_t *cluster, uint32_t value);
-attribute_t *create_uncertainty(cluster_t *cluster, uint16_t value);
+attribute_t *create_uncertainty(cluster_t *cluster, float value);
 attribute_t *create_measurement_unit(cluster_t *cluster, uint8_t value);
 attribute_t *create_measurement_medium(cluster_t *cluster, uint8_t value);
 attribute_t *create_level_value(cluster_t *cluster, uint8_t value);
@@ -575,14 +575,14 @@ attribute_t *create_level_value(cluster_t *cluster, uint8_t value);
 
 namespace pm25_concentration_measurement {
 namespace attribute {
-attribute_t *create_measured_value(cluster_t *cluster, nullable<uint16_t> value);
-attribute_t *create_min_measured_value(cluster_t *cluster, nullable<uint16_t> value);
-attribute_t *create_max_measured_value(cluster_t *cluster, nullable<uint16_t> value);
-attribute_t *create_peak_measured_value(cluster_t *cluster, nullable<uint16_t> value);
+attribute_t *create_measured_value(cluster_t *cluster, nullable<float> value);
+attribute_t *create_min_measured_value(cluster_t *cluster, nullable<float> value);
+attribute_t *create_max_measured_value(cluster_t *cluster, nullable<float> value);
+attribute_t *create_peak_measured_value(cluster_t *cluster, nullable<float> value);
 attribute_t *create_peak_measured_value_window(cluster_t *cluster, uint32_t value);
-attribute_t *create_average_measured_value(cluster_t *cluster, nullable<uint16_t> value);
+attribute_t *create_average_measured_value(cluster_t *cluster, nullable<float> value);
 attribute_t *create_average_measured_value_window(cluster_t *cluster, uint32_t value);
-attribute_t *create_uncertainty(cluster_t *cluster, uint16_t value);
+attribute_t *create_uncertainty(cluster_t *cluster, float value);
 attribute_t *create_measurement_unit(cluster_t *cluster, uint8_t value);
 attribute_t *create_measurement_medium(cluster_t *cluster, uint8_t value);
 attribute_t *create_level_value(cluster_t *cluster, uint8_t value);
@@ -591,14 +591,14 @@ attribute_t *create_level_value(cluster_t *cluster, uint8_t value);
 
 namespace pm10_concentration_measurement {
 namespace attribute {
-attribute_t *create_measured_value(cluster_t *cluster, nullable<uint16_t> value);
-attribute_t *create_min_measured_value(cluster_t *cluster, nullable<uint16_t> value);
-attribute_t *create_max_measured_value(cluster_t *cluster, nullable<uint16_t> value);
-attribute_t *create_peak_measured_value(cluster_t *cluster, nullable<uint16_t> value);
+attribute_t *create_measured_value(cluster_t *cluster, nullable<float> value);
+attribute_t *create_min_measured_value(cluster_t *cluster, nullable<float> value);
+attribute_t *create_max_measured_value(cluster_t *cluster, nullable<float> value);
+attribute_t *create_peak_measured_value(cluster_t *cluster, nullable<float> value);
 attribute_t *create_peak_measured_value_window(cluster_t *cluster, uint32_t value);
-attribute_t *create_average_measured_value(cluster_t *cluster, nullable<uint16_t> value);
+attribute_t *create_average_measured_value(cluster_t *cluster, nullable<float> value);
 attribute_t *create_average_measured_value_window(cluster_t *cluster, uint32_t value);
-attribute_t *create_uncertainty(cluster_t *cluster, uint16_t value);
+attribute_t *create_uncertainty(cluster_t *cluster, float value);
 attribute_t *create_measurement_unit(cluster_t *cluster, uint8_t value);
 attribute_t *create_measurement_medium(cluster_t *cluster, uint8_t value);
 attribute_t *create_level_value(cluster_t *cluster, uint8_t value);
@@ -607,14 +607,14 @@ attribute_t *create_level_value(cluster_t *cluster, uint8_t value);
 
 namespace radon_concentration_measurement {
 namespace attribute {
-attribute_t *create_measured_value(cluster_t *cluster, nullable<uint16_t> value);
-attribute_t *create_min_measured_value(cluster_t *cluster, nullable<uint16_t> value);
-attribute_t *create_max_measured_value(cluster_t *cluster, nullable<uint16_t> value);
-attribute_t *create_peak_measured_value(cluster_t *cluster, nullable<uint16_t> value);
+attribute_t *create_measured_value(cluster_t *cluster, nullable<float> value);
+attribute_t *create_min_measured_value(cluster_t *cluster, nullable<float> value);
+attribute_t *create_max_measured_value(cluster_t *cluster, nullable<float> value);
+attribute_t *create_peak_measured_value(cluster_t *cluster, nullable<float> value);
 attribute_t *create_peak_measured_value_window(cluster_t *cluster, uint32_t value);
-attribute_t *create_average_measured_value(cluster_t *cluster, nullable<uint16_t> value);
+attribute_t *create_average_measured_value(cluster_t *cluster, nullable<float> value);
 attribute_t *create_average_measured_value_window(cluster_t *cluster, uint32_t value);
-attribute_t *create_uncertainty(cluster_t *cluster, uint16_t value);
+attribute_t *create_uncertainty(cluster_t *cluster, float value);
 attribute_t *create_measurement_unit(cluster_t *cluster, uint8_t value);
 attribute_t *create_measurement_medium(cluster_t *cluster, uint8_t value);
 attribute_t *create_level_value(cluster_t *cluster, uint8_t value);
@@ -623,14 +623,14 @@ attribute_t *create_level_value(cluster_t *cluster, uint8_t value);
 
 namespace total_volatile_organic_compounds_concentration_measurement {
 namespace attribute {
-attribute_t *create_measured_value(cluster_t *cluster, nullable<uint16_t> value);
-attribute_t *create_min_measured_value(cluster_t *cluster, nullable<uint16_t> value);
-attribute_t *create_max_measured_value(cluster_t *cluster, nullable<uint16_t> value);
-attribute_t *create_peak_measured_value(cluster_t *cluster, nullable<uint16_t> value);
+attribute_t *create_measured_value(cluster_t *cluster, nullable<float> value);
+attribute_t *create_min_measured_value(cluster_t *cluster, nullable<float> value);
+attribute_t *create_max_measured_value(cluster_t *cluster, nullable<float> value);
+attribute_t *create_peak_measured_value(cluster_t *cluster, nullable<float> value);
 attribute_t *create_peak_measured_value_window(cluster_t *cluster, uint32_t value);
-attribute_t *create_average_measured_value(cluster_t *cluster, nullable<uint16_t> value);
+attribute_t *create_average_measured_value(cluster_t *cluster, nullable<float> value);
 attribute_t *create_average_measured_value_window(cluster_t *cluster, uint32_t value);
-attribute_t *create_uncertainty(cluster_t *cluster, uint16_t value);
+attribute_t *create_uncertainty(cluster_t *cluster, float value);
 attribute_t *create_measurement_unit(cluster_t *cluster, uint8_t value);
 attribute_t *create_measurement_medium(cluster_t *cluster, uint8_t value);
 attribute_t *create_level_value(cluster_t *cluster, uint8_t value);
